requirements.txt
absl-py==0.1.11
adium-theme-ubuntu==0.3.4
alembic==0.9.5
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.3.1
arrow==0.10.0
asn1crypto==0.24.0
astor==0.6.2
astroid==1.5.3
attrs==15.2.0
backports-abc==0.4
backports.functools-lru-cache==1.5
backports.ssl-match-hostname==3.4.0.2
backports.weakref==1.0.post1
beautifulsoup4==4.4.1
billiard==3.3.0.23
bleach==1.5.0
blinker==1.3
boto==2.48.0
celery==3.1.25
certifi==2017.7.27.1
cffi==1.11.4
chardet==3.0.4
click==6.7
configparser==3.5.0
contextlib2==0.5.5
cryptography==2.1.4
cycler==0.10.0
debtags==2.0
decorator==4.2.1
defer==1.0.6
diff-match-patch==20121119
dirspec==13.10
Django==1.8.3
django-ajaximage==0.2.8
django-apptemplates==0.0.1
django-braces==1.8.1
django-classy-tags==0.6.2
django-common-helpers==0.7.0
django-cors-headers==1.1.0
django-crispy-forms==1.4.0
django-cron==0.4.3
django-extensions==1.5.5
django-grappelli==2.7.1
django-grappelli-extensions==0.1.3
django-import-export==0.2.8
django-sslserver==0.15
djangorestframework==3.1.3
djangorestframework-camel-case==0.2.0
dnspython==1.15.0
duplicity==0.7.6
elasticsearch==5.4.0
email-validator==1.0.2
enum34==1.1.6
envparse==0.2.0
factory-boy==2.9.2
Faker==0.8.3
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Cache==0.13.1
Flask-Cors==3.0.3
Flask-Elasticsearch==0.2.5
Flask-JWT==0.3.2
Flask-Login==0.4.0
Flask-Migrate==2.1.1
Flask-Redis==0.3.0
Flask-Script==2.0.6
Flask-Scrypt==0.1.3.6
Flask-SQLAlchemy==2.1
forex-python==0.3.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.1.1
gast==0.2.0
gcm==0.3
geoip2==2.5.0
google-compute-engine==2.6.1
grpcio==1.10.0
gunicorn==19.7.1
healthcheck==1.3.2
html5lib==0.9999999
httplib2==0.9.1
humanize==0.5.1
icalendar==3.11.7
idna==2.6
ipaddress==1.0.19
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
isort==4.2.15
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
kiwisolver==1.0.1
kombu==3.0.37
lazy-object-proxy==1.3.1
librabbitmq==1.5.1
lockfile==0.12.2
lxml==3.5.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==0.23
marrow.mailer==4.0.2
marrow.util==1.2.3
marshmallow==2.13.1
marshmallow-jsonapi==0.15.1
matplotlib==2.2.0
maxminddb==1.3.0
mccabe==0.6.1
mistune==0.7.2
mock==2.0.0
nbconvert==4.1.0
nbformat==4.0.1
ndg-httpsclient==0.4.0
networkx==2.1
notebook==4.1.0
numpy==1.14.2
oauth2==1.9.0.post1
oauthlib==1.0.3
oneconf==0.3.9
PAM==0.4.2
path.py==8.1.2
pbr==3.1.1
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.6
Pillow==5.0.0
piston-mini-client==0.7.5
powerline-status==2.6
protobuf==3.5.2
psutil==3.4.2
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.7
PyAudio==0.2.8
pycountry==17.5.14
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.73
pygame===1.9.1release
Pygments==2.1.3
pygobject==3.20.0
PyJWT==1.4.2
pylint==1.7.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
PyPDF2==1.26.0
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.7.0
python-debian==0.1.27
python-editor==1.0.3
python-geoip==1.2
python-geoip-geolite2==2015.303
python-magic==0.4.13
python-pentabarf-xml==0.19
pytz==2018.3
PyWavelets==0.5.2
pyxdg==0.25
pyzmq==15.2.0
qrcode==5.3
qtconsole==4.2.1
raven==6.1.0
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
rope==0.10.7
scikit-image==0.13.1
scipy==1.0.0
scrypt==0.8.0
service-identity==16.0.0
simplegeneric==0.8.1
simplejson==3.11.1
sinchsms==1.0.4
singledispatch==3.4.0.3
six==1.11.0
software-center-aptd-plugins==0.0.0
SpeechRecognition==3.6.3
SQLAlchemy==1.1.14
SQLAlchemy-Continuum==1.3.1
SQLAlchemy-Utils==0.32.16
stripe==1.65.0
subprocess32==3.2.7
tablib==0.10.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
terminado==0.6
tornado==4.3
traitlets==4.2.1
twilio==5.6.0
Twisted==16.0.0
tzlocal==1.4
ubuntuone-client-data==14.4
unicode-slugify==0.1.3
Unidecode==0.4.21
unity-lens-photos==1.0
urllib3==1.22
vboxapi==1.0
virtualenv==15.0.1
Werkzeug==0.12.2
wrapt==1.10.10
WTForms==2.1
xhtml2pdf==0.0.6
zope.interface==4.1.3
